+++ CHANGE_LOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## [Unreleased]
+## [0.5.0] - 2026-03-23
 
 ## Added